Refactor Carousel (slide swiper) to a separate component that can be reused - we can now create a carousel on any page - added 'use strict' to react files
