+++ src/app/components/Header.js
@@ -1,3 +1,5 @@
+'use strict'
+
 import React, { Component } from 'react'
 import { NavLink } from 'react-router-dom'
 import { Dropdown, MenuItem } from 'react-bootstrap'
