+++ src/app/pages/IndexPage.js
@@ -1,55 +1,23 @@
+'use strict'
+
 import React, { Component } from 'react'
-import Swiper from 'react-id-swiper'
+import Carousel from '../components/Carousel'
 
 class IndexPage extends Component {
   render() {
 
-    document.title = "RED | Reforming Education on Drugs";
+    document.title = 'RED | Reforming Education on Drugs';
+    const imageList = ['main10.jpg', 'main3.jpg', 'main5.jpg', 'main8.jpg', 'main9.jpg']
 
     return(
       <main>
-        <SwiperSlider />
+        <Carousel images={imageList}/>
         <Description />
         {/* TODO: Testimonials */}
         {/* TODO: MailChimp Sign Up form */}
         {/* TODO: Brands */}
         {/* TODO: Footer */}
       </main>
-    )
-  }
-}
-
-class SwiperSlider extends Component {
-  render() {
-
-    const swiperParams = {
-      slidesPerView: 1,
-      autoplay: {
-        delay: 5000,
-      },
-      loop: true,
-      pagination: {
-        el: '.swiper-pagination',
-        clickable: true,
-      },
-      navigation: {
-        nextEl: '.swiper-button-next',
-        prevEl: '.swiper-button-prev',
-      },
-    }
-
-    return (
-      <div className="container" id="swiper">
-        <div className="simple-slider">
-          <Swiper {...swiperParams}>
-            <div style={{backgroundImage: 'url(/assets/img/main10.jpg)'}} />
-            <div style={{backgroundImage: 'url(/assets/img/main3.jpg)'}} />
-            <div style={{backgroundImage: 'url(/assets/img/main5.jpg)'}} />
-            <div style={{backgroundImage: 'url(/assets/img/main8.jpg)'}} />
-            <div style={{backgroundImage: 'url(/assets/img/main9.jpg)'}} />
-          </Swiper>
-        </div>
-      </div>
     )
   }
 }
@@ -62,7 +30,7 @@ class Description extends Component {
         <div className="container">
           <h1 className="text-center">Reforming Education on Drugs</h1>
           <p>RED’s mission is to establish within students a foundational understanding of the biological mechanisms involved in illicit drug use and substance abuse. Through the use of interactive in-class presentations and demonstrations, we provide individuals with the ability to engage and think critically with these topics, all the while developing a love for science learning.</p>
-        </div>
+        </div>  
       </div>
     )
   }
