+++ src/app/pages/NotFoundPage.js
@@ -1,3 +1,5 @@
+'use strict'
+
 import React, { Component } from 'react'
 
 class NotFoundPage extends Component {
